Uncomment Playwright testing to start implementing CI
Making Playwright work requires setting up an environment on GitHub.
Multiple commits will be made to help set the workflow file to a state
that allows for the web app to be hosted and tested. For now, I've
listed a single environment variable, but more must be added - probably
with a database running in the background too.